% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_groups_to_xlsform_names.R
\name{add_groups_to_xlsform_names}
\alias{add_groups_to_xlsform_names}
\title{Add group names to XLSForm survey sheet name column}
\usage{
add_groups_to_xlsform_names(survey, sep = ".")
}
\arguments{
\item{survey}{XLSForm survey sheet}

\item{sep}{Separator for groups from name of question, either `.` or `/`}
}
\value{
XLSForm survey sheet with edited name column
}
\description{
Add group names to XLSForm survey sheet name column
}
