% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detect_select_multiple.R
\name{auto_detect_select_multiple}
\alias{auto_detect_select_multiple}
\title{Auto Detect Select Multiple Questions}
\usage{
auto_detect_select_multiple(df)
}
\arguments{
\item{df}{dataset (data frame)}
}
\value{
vector of character column names detected as select multiple
}
\description{
Auto Detect Select Multiple Questions
}
