% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detects.R
\name{auto_detect_sm_parents}
\alias{auto_detect_sm_parents}
\title{Detect select multiple parent columns}
\usage{
auto_detect_sm_parents(df, sm_sep = "/")
}
\arguments{
\item{df}{a survey object or dataframe}

\item{sm_sep}{select multiple parent child separator. This is specific for XLSForm data (default = /).
If using read_csv to read in data the separator will most likely be '/' where as if using read.csv it will likely be '.'}
}
\value{
a list of select multiple parent columns in data set.
}
\description{
`auto_detect_sm_parents` is mean to detect select multiple parent columns in a way that does
not rely on the XLSForm as the input
}
