% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_detects.R
\name{auto_sm_parent_child}
\alias{auto_sm_parent_child}
\title{detect and group together select multiple parent and children columns}
\usage{
auto_sm_parent_child(df, sm_sep = "/")
}
\arguments{
\item{df}{a survey object or dataframe}

\item{sm_sep}{select multiple parent child separator. This is specific for XLSForm data (default = /).
If using read_csv to read in data the separator will most likely be '/' where as if using read.csv it will likely be '.'}
}
\value{
a data frame containing the the child select multiple columns alongside there parents
}
\description{
`auto_sm_parent_child` is mean to detect select multiple parent columns & children columns in a way that does
not rely on the XLSForm as the input
}
