% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_by_groups.R
\name{barplot_by_group}
\alias{barplot_by_group}
\title{Barplot grouped binary columns}
\usage{
barplot_by_group(design, list_of_variables, aggregation_level, binary = TRUE)
}
\arguments{
\item{design}{Design object from survey or srvyr package.}

\item{list_of_variables}{Vector containing column names to analyze.}

\item{aggregation_level}{Column name to aggregate or dissagregate to OR vector of column names to dissagregate to.}

\item{binary}{Logical (default=TRUE) if the columns are binary or numeric.}
}
\description{
Barplot grouped binary columns
}
