% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_grouped_binaries.R
\name{barplot_grouped_binaries}
\alias{barplot_grouped_binaries}
\title{BARPLOT GROUPED BINARY COLUMNS}
\usage{
barplot_grouped_binaries(design, list_of_variables, aggregation_level)
}
\arguments{
\item{design}{design object}

\item{list_of_variables}{list of variables that you want to plot together}

\item{aggregation_level}{the level to aggregate calculations to (this will ultimately be categorical on the Y-Axis)}
}
\description{
This function takes a list of variables from a survey object and puts them all on one barplot.
}
