% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_cleaning_log.R
\name{check_cleaning_log}
\alias{check_cleaning_log}
\title{check cleaning log}
\usage{
check_cleaning_log(
  df,
  df_uuid,
  cl,
  cl_change_type_col,
  cl_change_col,
  cl_uuid,
  cl_new_val
)
}
\arguments{
\item{df}{raw data (data.frame)}

\item{df_uuid}{column in raw data with uuid}

\item{cl}{cleaning log (data.frame)}

\item{cl_change_type_col}{column in cleaning log which specifies change type to be made}

\item{cl_change_col}{column in cleaning log which specifies data set column to change}

\item{cl_uuid}{uuid in cleaning log}

\item{cl_new_val}{cleaning log column specifying the new correct value}
}
\value{
cleaning log with only problematic entries and note specifying problem
}
\description{
check cleaning log
}
