% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distances_from_target.R
\name{check_distances_from_target}
\alias{check_distances_from_target}
\title{check distance from target}
\usage{
check_distances_from_target(
  dataset_sf,
  target_points,
  cols_to_report,
  distance_threshold = 25
)
}
\arguments{
\item{target_points}{spatial points to check dataset agains}

\item{cols_to_report}{which columns you want to return with distnace}

\item{dataset}{an sf dataset}
}
\value{
simplified dataset with distance to closest target point
}
\description{
check distance from target
}
