% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_main_repeat_consistency.R
\name{check_main_repeat_consistency}
\alias{check_main_repeat_consistency}
\title{Check consistencies between main data set and repeat data set}
\usage{
check_main_repeat_consistency(main_dataset, main_col1, main_col2,
  main_uuid, repeat_dataset, repeat_col1, repeat_col2, repeat_uuid)
}
\arguments{
\item{main_dataset}{main data set}

\item{main_col1}{column 1 in main data}

\item{main_col2}{column 2 in main data}

\item{main_uuid}{unique identifier in main data set to match unique identifier in repeat}

\item{repeat_dataset}{repeat data set}

\item{repeat_col1}{column 1 in repeat data}

\item{repeat_col2}{column 2 in repeat data}

\item{repeat_uuid}{unique identifier in repeat data set to match unique identifier in repeat}
}
\value{
uuids of non consistent results
}
\description{
Check consistencies between main data set and repeat data set
}
