% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_others.R
\name{check_others}
\alias{check_others}
\title{Check variation by: check the variation of all columns by grouping var (specifically designed for the grouping var to be the data collectors identifier). Variation results are checked with resepect to standard deviations of them mean and a plot is produced to help understand issues.}
\usage{
check_others(df, suffix = "_other", report_cols)
}
\arguments{
\item{df}{data frame}

\item{suffix}{suffix that ends other columns. others need systematic suffix (i.e "_other")}

\item{report_cols}{columns to report}
}
\value{
a useful table of other values and a cleaning log
}
\description{
Check variation by: check the variation of all columns by grouping var (specifically designed for the grouping var to be the data collectors identifier). Variation results are checked with resepect to standard deviations of them mean and a plot is produced to help understand issues.
}
