% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_point_distances_by_id.R
\name{check_point_distance_by_id}
\alias{check_point_distance_by_id}
\title{check distances between points by id}
\usage{
check_point_distance_by_id(sf1, sf2, sf1_id, sf2_id, dist_threshold)
}
\arguments{
\item{sf1}{sf class point class object 1}

\item{sf2}{sf class point class object 2}

\item{sf1_id}{sf1_id that links to sf_2}

\item{sf2_id}{sf2_id that links to sf_1}

\item{dist_threshold}{distance threshold to check}
}
\value{
1.) dataset with distance column mutated, 2.) a leaflet map, 3.) histogram of distances
}
\description{
check distances between points by id
}
