% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reported_strata_against_spatial_poly.R
\name{check_reported_strata_against_spatial_poly}
\alias{check_reported_strata_against_spatial_poly}
\title{check_reported_strata_against_spatial_poly}
\usage{
check_reported_strata_against_spatial_poly(
  dataset,
  strata_poly,
  dataset_coordinates = c("X_gps_reading_longitude", "X_gps_reading_latitude"),
  dataset_strata_name,
  poly_strata_name,
  cols_to_report = c("X_uuid", "enumerator_id")
)
}
\arguments{
\item{dataset}{data set with coordinates}

\item{strata_poly}{polygon (converted to sf object) which defines strata (administrative unit, camp boundary, etc.)}

\item{dataset_coordinates}{columns in dataset with hold longitude latitude in that order. default is c("X_gps_reading_longitude" , "X_gps_reading_latitude")}

\item{dataset_strata_name}{column name which hold the strata in the datset}

\item{poly_strata_name}{column which hold the strata in the polygon file}

\item{cols_to_report}{columns which you want to report along with mismatched strata. default is c("X_uuid","enumerator_id")}
}
\value{
dataframe where strata reported in dataset do not match strata in spatial admin file according to  coordinates in data set
}
\description{
check_reported_strata_against_spatial_poly
}
