% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variation_by.R
\name{check_variation_by}
\alias{check_variation_by}
\title{Check variation by: check the variation of all columns by grouping var (specifically designed for the grouping var to be the data collectors identifier). Variation results are checked with resepect to standard deviations of them mean and a plot is produced to help understand issues.}
\usage{
check_variation_by(df, by = "enumerator_id", zscore_threshold = 3)
}
\arguments{
\item{df}{data frame}

\item{by}{What you want to check by (designed specifically with enumerator in mind)}

\item{zscore_threshold}{How many standard deviations to consider outlier (default 3)}
}
\value{
ggplot bar graph showing top problematic groups (enumerators), table of issues
}
\description{
Check variation by: check the variation of all columns by grouping var (specifically designed for the grouping var to be the data collectors identifier). Variation results are checked with resepect to standard deviations of them mean and a plot is produced to help understand issues.
}
