% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_closest_pt_rtree.R
\name{closest_distance_rtree}
\alias{closest_distance_rtree}
\title{closest_distance_rtree}
\usage{
closest_distance_rtree(sf1, sf2, k = 1, sf1_coords = c("X", "Y"))
}
\arguments{
\item{sf1}{simple feature 1 (point)}

\item{sf2}{simple feature 2 (point).}
}
\value{
sf1 with closest record from sf2 attached.
}
\description{
closest_distance_rtree
}
