% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_file_prefix.R
\name{date_file_prefix}
\alias{date_file_prefix}
\title{Use date string to prefix file names}
\usage{
date_file_prefix(days_from_present = 0)
}
\arguments{
\item{days_from_present}{number of days from present. Default is today 0. Tomorrow would be 1, yesterday= -1}
}
\value{
date formated ISO86031 no space
}
\description{
Use date string to prefix file names
}
