% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_comparison_log.R
\name{df_comparison_log}
\alias{df_comparison_log}
\title{Generate comparison/cleaning log between a raw/unchanged data set and a modified/cleaned data set}
\usage{
df_comparison_log(raw_data, clean_data, raw_data_uuid, clean_data_uuid)
}
\arguments{
\item{raw_data}{raw/unchanged data set (df1)}

\item{clean_data}{clean/changed data set (df2)}

\item{raw_data_uuid}{name of unique identifier column in df1 that matches values of unique identifier in df2}

\item{clean_data_uuid}{name of unique identifier column in df2 that matches values of unique identifier in df1}
}
\value{
a comparison/cleaning log that shows all changes that were made from teh raw data set
}
\description{
Generate comparison/cleaning log between a raw/unchanged data set and a modified/cleaned data set
}
