% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sm_option_columns.R
\name{extract_sm_option_columns}
\alias{extract_sm_option_columns}
\title{Find and replace concatenated select multiple column with all individual select multiple options}
\usage{
extract_sm_option_columns(df, name_vector)
}
\arguments{
\item{df}{data frame}

\item{name_vector}{vector containing column names in the data set.}

\item{aggregation_level}{Column name to aggregate or dissagregate to OR vector of column names to dissagregate to.}
}
\value{
dataframe with parent  (concatenate select mutiple column) and individual select multiple option columns
}
\description{
Find and replace concatenated select multiple column with all individual select multiple options
}
