% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_response_rates.R
\name{get_na_response_rates}
\alias{get_na_response_rates}
\title{GET NA PERCENTAGES AND FREQUENCY FOR EACH COLUMN IN DATA SET}
\usage{
get_na_response_rates(data)
}
\arguments{
\item{data}{the dataset}
}
\description{
returns the count of NA and the percentage of NA for each column in a dataset
}
