% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_audit.R
\name{load_audit}
\alias{load_audit}
\title{Load audit files}
\usage{
load_audit(
  data,
  path.to.zip,
  path.to.unzip,
  copy.zip = TRUE,
  path.to.copy.zip,
  filter.column = "informed_consent",
  filter.on = "yes",
  uuid.column = "X_uuid",
  delete.unzipped = TRUE,
  days.ago.reported = 0
)
}
\arguments{
\item{data}{data set}

\item{path.to.zip}{where audit zip file is stored}

\item{copy.zip}{logical whether or not to copy zip (logical = TRUE, yes copy zip)}

\item{path.to.copy.zip}{where you want to copy zip (dropbox)}

\item{filter.column}{column to filter data set (informed consent usually)}

\item{filter.on}{filter value to remove}

\item{uuid.column}{column contining unique identifier}

\item{delete.unzipped}{logical- delete unzipped file (TRUE default= delete)}

\item{days.ago.reported}{days since data was reported (default = 0)}
}
\value{
audit data
}
\description{
Load audit files
}
