% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_functions.R
\name{lookup_to_data_validation}
\alias{lookup_to_data_validation}
\title{lookup table to data validation}
\usage{
lookup_to_data_validation(lookup_table)
}
\arguments{
\item{lookup_table}{output form xlsform_lookuptalbe}
}
\value{
data frame that can be added as xlsx sheet to easily make data validation for kobo tool/cleaning log
}
\description{
lookup table to data validation
}
