% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_functions.R
\name{make_xlsform_lookup_table}
\alias{make_xlsform_lookup_table}
\title{Make XlsForm lookup table from tool}
\usage{
make_xlsform_lookup_table(kobo_survey, kobo_choices, label_column)
}
\arguments{
\item{kobo_survey}{survey sheet from tool}

\item{kobo_choices}{choices sheet from tool}

\item{label_column}{label column used in tool}
}
\value{
will return a lookup table with columns containing mergeable columns with the data set as well as butteR/survey package analyzed dataset
}
\description{
Make XlsForm lookup table from tool
}
