% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_closest_pt_rtree.R
\name{match_closest_pt_rtree}
\alias{match_closest_pt_rtree}
\title{Find closest point between to spatial point objects using rtree spatial indexing}
\usage{
match_closest_pt_rtree(sf1, sf2, return_distance = TRUE)
}
\arguments{
\item{sf1}{simple feature 1 (point)}

\item{sf2}{simple feature 2 (point).}

\item{return_distance}{return the closest point with a vector indicating distance (m). Default = TRUE}
}
\value{
sf1 with closest record from sf2 attached.
}
\description{
Find closest point between to spatial point objects using rtree spatial indexing
}
