% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_functions.R
\name{survey_name_choice_name_match}
\alias{survey_name_choice_name_match}
\title{Mutate multiple columns at once with same value}
\usage{
survey_name_choice_name_match(kobold)
}
\arguments{
\item{kobold}{kobold object}
}
\description{
mutating batch columns allows programmic creation of columns based
on an input vector or list. This was developed to add on new variables to a xlsform data
set which can then be systematically added to the the xlsform tool
}
