% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_prop_working.R
\name{mean_prop_working}
\alias{mean_prop_working}
\title{Generate Summary Tables For Chosen Columns/Indicators/Questions}
\usage{
mean_prop_working(design, list_of_variables, aggregation_level = NULL,
  round_to = 2, return_confidence = TRUE, na_replace = FALSE)
}
\arguments{
\item{design}{Design object.}

\item{list_of_variables}{Vector containing all variable names to analyze and include in summary table.}

\item{aggregation_level}{Aggregation level(s) to analyze data by. The default (NULL) will fully aggregate data to provide one value per variable. Argument also accepts vecors to aggregate the data using mroe than one variable.}

\item{round_to}{The number of digits to round reults to.}

\item{return_confidence}{Logical variable. TRUE (default) will return 95 percent confidence interval, FALSE returns no confidence interval.}

\item{na_replace}{Logical variable. TRUE will replace NA with 0 for integerss and "filtered value" for categorical variables. FALSE (default) will leave NAs in dataset and thus they will automatically be removed during calculation.}
}
\value{
Table of summarised, aggregated variables.
}
\description{
Add all levels from possible answer choices from choices sheet to dataset. This reduces
possible errors that can occur when aggregating and computing summary statistics with survey package
}
