% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_proportion_table.R
\name{mean_proportion_table}
\alias{mean_proportion_table}
\title{Analyze and aggregate weighted categorical data into clean tables}
\usage{
mean_proportion_table(design, list_of_variables,
  aggregation_level = NULL, round_to = 2, return_confidence = TRUE,
  na_replace = FALSE, questionnaire = NULL)
}
\arguments{
\item{design}{Design object from survey or srvyr package.}

\item{list_of_variables}{Vector containing column names to analyze.}

\item{aggregation_level}{Column name to aggregate or dissagregate to OR vector of column names to dissagregate to.}

\item{round_to}{Decimal place to round to.}

\item{return_confidence}{Logical value specifying whether to return confidence interval.}

\item{na_replace}{Logical value (default = FALSE) of whether to replace NA with 0 (numerical) or "filtered" (categorical)}

\item{questionnaire}{Questionnaire generated from koboquest. If NULL (default) function will attempt to detect select multiple questions automatically. If you have made new indicators which use periods in there names you must remove them from the list_of_variables if you do not supply the koboquest generated questionnaire.}
}
\value{
Analyzed table of variables.
}
\description{
Analyze and aggregate weighted categorical data into clean tables
}
