% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_functions.R
\name{mutate_batch}
\alias{mutate_batch}
\title{Mutate multiple columns at once with same value}
\usage{
mutate_batch(df, nm, value = NA)
}
\arguments{
\item{df}{dataframe}

\item{names}{names of columns to mutate}

\item{values}{uniform values to mutate}
}
\description{
mutating batch columns allows programmic creation of columns based
on an input vector or list. This was developed to add on new variables to a xlsform data
set which can then be systematically added to the the xlsform tool
}
