% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_collapse.R
\name{mutate_key_pair}
\alias{mutate_key_pair}
\title{Mutate columns on based on a list of names and values}
\usage{
mutate_key_pair(df, names, values)
}
\arguments{
\item{df}{dataframe}

\item{names}{names of columns to mutate}

\item{values}{uniform values to mutate}
}
\description{
conditionally mutate on columns based
on a list of column names and values. This is mostly useful for conditional
mutate commands and can currently only mutate uniform columns.
It is used inside the survey collapse functions
}
