% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_nearest_feature.R
\name{mutate_nearest_feature}
\alias{mutate_nearest_feature}
\title{mutate_nearest_feature}
\usage{
mutate_nearest_feature(x, y)
}
\arguments{
\item{x}{sf object to measure by element closest distance to all y features}

\item{y}{sf object to query against each row feature in x}
}
\value{
x with the distance to the closest feature from y (distance) and its index (y_index) mutated as
two columns of the data set
}
\description{
convenience wrapper for st_nearest_feature & st_distance
}
