% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_others.R
\name{others_cleaning_log}
\alias{others_cleaning_log}
\title{others_cleaning_log}
\usage{
others_cleaning_log(df, suffix = "_other", report_cols, kobo_survey_sheet)
}
\arguments{
\item{df}{data.frame kobo/odk/xlsform style questionnaire respose data as data.frame}

\item{suffix}{identifying suffix for "other" free-text columns (default:  "_other"-recommended tool design)}

\item{report_cols}{columns to report (must include uuid)}

\item{kobo_survey_sheet}{data.frame containing kobo/xlsForm survey tab}
}
\value{
semi-automated cleaning log formatted for use with kobold::kobold_cleaner. This log should then
be manually reviewed.
}
\description{
generate semi-auto cleaning log for regrouping and "other" free-text responses
}
