% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variation_by.R
\name{outliers_by_to_cleaning_log_template}
\alias{outliers_by_to_cleaning_log_template}
\title{outliers_by_to_cleaning_log_template}
\usage{
outliers_by_to_cleaning_log_template(outliers_by = low_outliers, by = by)
}
\arguments{
\item{outliers_by}{data frame indicating outliers grouped by chosen var}

\item{by}{What you want to check by (designed specifically with enumerator in mind)}
}
\value{
cleaning log template
}
\description{
outliers_by_to_cleaning_log_template
}
