% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_others.R
\name{parent_other_q}
\alias{parent_other_q}
\title{parent_other_q}
\usage{
parent_other_q(df)
}
\arguments{
\item{df}{kobo/odk/xlsform style questionnaire respose data as data.frame}
}
\value{
a long format data frame containing the collapsed data.
}
\description{
For questions that have other free-text follow up option, get parent and child variables
}
