% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change_calculations.R
\name{pct_change}
\alias{pct_change}
\title{Percent change of a variable compared to previous}
\usage{
pct_change(x)
}
\arguments{
\item{x}{variable name}
}
\description{
calculate percent change for a particular numeric variable compared to a previous time-step
}
