% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change_calculations.R
\name{pct_change_by_groups_all_numerics}
\alias{pct_change_by_groups_all_numerics}
\title{Percent change by groups for all numerics in dataframe}
\usage{
pct_change_by_groups_all_numerics(df, group_var, time_id)
}
\arguments{
\item{df}{data frame}

\item{group_var}{the variable to group_by. This column will be reported in final data.}

\item{time_id}{column name containing the time_id (i.e month, year, etc.)}
}
\value{
data.frame with grouping var and all the same columns but half the number of rows
containing the % change for each value.
}
\description{
Used to calculate percent change of all numeric data by a group/strata level.
Created specifically for calculating price % change for many items across months
}
