% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_factorize_categorical.R
\name{questionnaire_factorize_categorical}
\alias{questionnaire_factorize_categorical}
\title{Factorize all categorical data in data set according to questionnaire options.}
\usage{
questionnaire_factorize_categorical(
  data,
  questionnaire,
  return_full_data = FALSE
)
}
\arguments{
\item{data}{the data set}

\item{questionnaire}{Questionnaire genereated from kobo quest package.}

\item{return_full_data}{Logical (default=FALSE). If TRUE refactored data will be inserted back into original dataset. If FALSE (default), only the refactored data will be returned.}

\item{round_to}{Decimal place to round to.}

\item{return_confidence}{Logical value specifying whether to return confidence interval.}

\item{na_replace}{Logical value (default = FALSE) of whether to replace NA with 0 (numerical) or "filtered" (categorical)}
}
\value{
Dataset with cateogrical data factorized according to questionnaire.
}
\description{
Factorize all categorical data in data set according to questionnaire options.
}
