% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_csvs_in_folder.R
\name{readr_all_csvs_in_folder}
\alias{readr_all_csvs_in_folder}
\title{readr_all_csvs_in_folder}
\usage{
readr_all_csvs_in_folder(input_csv_folder)
}
\arguments{
\item{input_csv_folder}{folder containing csvs of interest}
}
\value{
a list of data frames, each one named after the file name (uses readr)
}
\description{
readr_all_csvs_in_folder
}
