% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_functions.R
\name{refactor_to_xlsform}
\alias{refactor_to_xlsform}
\title{refactor to xlsform}
\usage{
refactor_to_xlsform(
  data,
  kobo_survey,
  kobo_choices,
  label_column = "label::english"
)
}
\arguments{
\item{data}{data set}

\item{kobo_choices}{choices sheet from tool}

\item{label_column}{label column used in tool}
}
\value{
data set with select one questions refactored to kobo tool
}
\description{
refactor to xlsform
}
