% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_concat_select_multiple.R
\name{remove_concat_select_multiple}
\alias{remove_concat_select_multiple}
\title{REMOVE CONCATENATED SELECT MULTIPLE COLUMNS FROM DATA SET}
\usage{
remove_concat_select_multiple(data, questionnaire)
}
\arguments{
\item{data}{the dataset (data frame). Does not yet support design objects}

\item{questionnaire}{questionnaire object created by the koboquest package}
}
\description{
returns the input data frame with concatenated select multiple questions
removed
}
