% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_groups_from_xlsform_names.R
\name{remove_groups_from_xlsform_names}
\alias{remove_groups_from_xlsform_names}
\title{Remove group names from XLSForm survey sheet name column}
\usage{
remove_groups_from_xlsform_names(survey, sep = ".")
}
\arguments{
\item{survey}{XLSForm survey sheet}

\item{sep}{Separator for groups from name of question, either `.` or `/`}
}
\value{
XLSForm survey sheet with original name column
}
\description{
Works to remove group names added from the `add_group_to_xlsform_name` function to return the original survey sheet.
}
