% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_kobo_grouper}
\alias{remove_kobo_grouper}
\title{Remove grouper added to kobo dataset when downloaded}
\usage{
remove_kobo_grouper(colname, max_prefix_length)
}
\arguments{
\item{colname}{colname vector}
}
\value{
colname vector without grouper
}
\description{
Remove grouper added to kobo dataset when downloaded
}
