% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_select_multiple_groups.R
\name{return_select_multiple_groups}
\alias{return_select_multiple_groups}
\title{RETURN LIST OF DATA FRAMES FOR EACH SELECT MULTIPLE
The data frame will be split and grouped by each select multiple question}
\usage{
return_select_multiple_groups(data, questionnaire)
}
\arguments{
\item{data}{the data set}

\item{questionnaire}{the questionnaire object generated by koboquest}
}
\description{
RETURN LIST OF DATA FRAMES FOR EACH SELECT MULTIPLE
The data frame will be split and grouped by each select multiple question
}
