% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_drop_geometry_keep_coords.R
\name{st_drop_geometry_keep_coords}
\alias{st_drop_geometry_keep_coords}
\title{Wraps SF st_drop_geometry but retains coordinates as first column in data.frame}
\usage{
st_drop_geometry_keep_coords(sf_object)
}
\arguments{
\item{sf_objec}{simple_feature object}
}
\value{
sf object as data frame with coordinates as first column
}
\description{
Wraps SF st_drop_geometry but retains coordinates as first column in data.frame
}
