% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_sampler.R
\name{stratified_sampler}
\alias{stratified_sampler}
\title{Stratified simple random sampling.}
\usage{
stratified_sampler(
  sample.target.frame,
  sample.target.frame.strata,
  sample.target.frame.samp.size,
  pt.data,
  pt.data.strata,
  pt.data.labels,
  target_gdb = NULL,
  seed = NULL,
  write_kml = TRUE,
  write_remaining_sample_csv = TRUE
)
}
\arguments{
\item{sample.target.frame}{data frame with strata, and sample size columns}

\item{sample.target.frame.strata}{name of strata column in sample.target.frame}

\item{sample.target.frame.samp.size}{name of sample size column in sample.target.frame}

\item{pt.data}{pt data object of class}

\item{pt.data.strata}{name of strata column in pt.data}

\item{pt.data.labels}{name of column to label individual kml points}

\item{target_gdb}{folde to output kml files  to. If empty/null (default) files will write to current working directory}

\item{seed}{set and save random seed (will set by default)}

\item{write_kml}{logical, write output to kml}

\item{write_remaining_sample_kml}{logical, write remaining sample to csv}
}
\value{
list with each sample as sf object. Exports each item in list to kml file in selected directory
}
\description{
Stratified simple random sampling.
}
