% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_on_multiple.R
\name{strsplit_on_multiple}
\alias{strsplit_on_multiple}
\title{Split strings on multiple delimeters}
\usage{
strsplit_on_multiple(x, separator)
}
\arguments{
\item{x}{string or vector of strings}

\item{separator}{single or vector of separator delimeters}
}
\value{
a list of vectors with split strings
}
\description{
Split strings on multiple delimeters
}
\details{
does not support regex matching
}
\examples{
strings_to_split<-c('abc def.gh', "def.kl abc 9")
separation_pattern<-c(" ", "c")
strsplit_on_multiple(x = strings_to_split,
                     separation_pattern)


}
