% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolower_rm_special.R
\name{tolower_rm_special}
\alias{tolower_rm_special}
\title{tolower_rm_special}
\usage{
tolower_rm_special(string)
}
\arguments{
\item{string}{character string}
}
\value{
lower case string all special characters and white space removed
}
\description{
tolower_rm_special
}
