% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_functions.R
\name{xlsform_add_choices}
\alias{xlsform_add_choices}
\title{Mutate multiple columns at once with same value}
\usage{
xlsform_add_choices(kobold, new_choices)
}
\arguments{
\item{new_choices}{new choice sheet containing question name and new choices}

\item{kobld}{kobold object}
}
\description{
mutating batch columns allows programmic creation of columns based
on an input vector or list. This was developed to add on new variables to a xlsform data
set which can then be systematically added to the the xlsform tool
}
