% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_by_groups.R
\name{barplot_by_group}
\alias{barplot_by_group}
\title{BARPLOT GROUPED BINARY COLUMNS}
\usage{
barplot_by_group(design, list_of_variables, aggregation_level,
  binary = TRUE)
}
\arguments{
\item{design}{design object}

\item{list_of_variables}{list of variables that you want to plot together}

\item{aggregation_level}{the level to aggregate calculations to (this will ultimately be categorical on the Y-Axis)}

\item{binary}{If the variables are binary or numeric}
}
\description{
This function takes a list of variables from a survey object and puts them all on one barplot.
}
