% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_grouped_binaries.R
\name{barplot_grouped_binaries}
\alias{barplot_grouped_binaries}
\title{Barplot grouped binary columns}
\usage{
barplot_grouped_binaries(design, list_of_variables, aggregation_level)
}
\arguments{
\item{design}{Design object from survey or srvyr package.}

\item{list_of_variables}{Vector containing column names to analyze.}

\item{aggregation_level}{Column name to aggregate or dissagregate to OR vector of column names to dissagregate to.}
}
\description{
Barplot grouped binary columns
}
