% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distances_from_target.R
\name{check_distances_from_target}
\alias{check_distances_from_target}
\title{check distance from target}
\usage{
check_distances_from_target(dataset, target_points, dataset_coordinates,
  cols_to_report, distance_threshold = 25)
}
\arguments{
\item{dataset}{dataset}

\item{target_points}{spatial points to check data set agains}

\item{dataset_coordinates}{column names for coordinates in dataset}

\item{cols_to_report}{which columns you want to return with distnace}
}
\value{
simplified dataset with distance to closest target point
}
\description{
check distance from target
}
