% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_file_prefix.R
\name{date_file_prefix}
\alias{date_file_prefix}
\title{Use date string to prefix file names}
\usage{
date_file_prefix(date = Sys.Date())
}
\arguments{
\item{date}{date object (default = Sys.Date())}
}
\value{
date formated ISO86031 no space
}
\description{
Use date string to prefix file names
}
