% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_factorize_categorical.R
\name{questionnaire_factorize_categorical}
\alias{questionnaire_factorize_categorical}
\title{FACTORIZE ALL CATEGORICAL DATA IN DATA FRAME ACCORDING TO KOBO/ODK QUESTIONNAIRE}
\usage{
questionnaire_factorize_categorical(data, questionnaire,
  return_full_data = FALSE)
}
\arguments{
\item{data}{the data set}

\item{questionnaire}{koboquest generated questionnnaire object}

\item{return_full_data}{logical. If true, refactored data will be inserted back into original dataset. If FALSE (default), only the refactored data will be returned}
}
\description{
Add all levels from possible answer choices from choices sheet to dataset. This reduces
possible errors that can occur when aggregating and computing summary statistics with survey package
}
