% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_concat_select_multiple.R
\name{remove_concat_select_multiple}
\alias{remove_concat_select_multiple}
\title{Remove concatenated select multiple columns from data set}
\usage{
remove_concat_select_multiple(data, questionnaire)
}
\arguments{
\item{data}{The dataset (data frame).}

\item{questionnaire}{questionnaire object created by the koboquest package}
}
\value{
data set without concatenated select multiple columns
}
\description{
Remove concatenated select multiple columns from data set
}
